Extract fake data types into types file
The `FakeTestData` class was getting overcrowded with type definitions.
